chore: move PR tests into src/react/tests/
Consolidate all test files touched by this PR into a single
src/react/tests/ directory so they don't clutter the source tree:

  tests/audio-element.test.ts    (was src/react/audio-element.test.ts)
  tests/jsx-async.test.tsx       (was src/react/jsx-async.test.tsx)
  tests/captions.test.ts         (was src/react/renderers/captions.test.ts)
  tests/ir-compile.test.ts       (was src/react/ir/tests/compile.test.ts)
  tests/ir-execute.test.ts       (was src/react/ir/tests/execute.test.ts)

IR tests renamed with ir- prefix for grouping. Imports updated
to reflect the new depth (one level deeper).
